requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --no-emit-index-url --strip-extras
#
anyio==4.9.0
    # via httpx
appdirs==1.4.4
    # via twisted
attrs==25.3.0
    # via
    #   hypothesis
    #   service-identity
    #   twisted
automat==24.8.1
    # via twisted
bcrypt==4.3.0
    # via twisted
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
constantly==23.10.4
    # via twisted
cryptography==44.0.2
    # via
    #   pyopenssl
    #   service-identity
    #   twisted
cython-test-exception-raiser==1.0.2
    # via twisted
datetype==2025.2.13
    # via quickmacapp (pyproject.toml)
h11==0.14.0
    # via httpcore
h2==4.2.0
    # via
    #   httpx
    #   twisted
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via twisted
hyperframe==6.1.0
    # via h2
hyperlink==21.0.0
    # via twisted
hypothesis==6.130.8
    # via twisted
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   twisted
incremental==24.7.2
    # via twisted
priority==1.3.0
    # via twisted
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.22
    # via cffi
pyhamcrest==2.1.0
    # via twisted
pyobjc-core==11.0
    # via
    #   pyobjc-framework-cfnetwork
    #   pyobjc-framework-cocoa
    #   pyobjc-framework-exceptionhandling
    #   pyobjc-framework-usernotifications
    #   twisted
pyobjc-framework-cfnetwork==11.0
    # via twisted
pyobjc-framework-cocoa==11.0
    # via
    #   pyobjc-framework-cfnetwork
    #   pyobjc-framework-exceptionhandling
    #   pyobjc-framework-usernotifications
    #   quickmacapp (pyproject.toml)
    #   twisted
pyobjc-framework-exceptionhandling==11.0
    # via quickmacapp (pyproject.toml)
pyobjc-framework-usernotifications==11.0
    # via quickmacapp (pyproject.toml)
pyopenssl==25.0.0
    # via twisted
pyserial==3.5
    # via twisted
service-identity==24.2.0
    # via twisted
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via hypothesis
twisted==24.11.0
    # via quickmacapp (pyproject.toml)
typing-extensions==4.13.1
    # via twisted
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
